+++ js/data.js
@@ -1,0 +1,67 @@
+const descriptions = [
+  'Прекрасный момент',
+  'Незабываемый день',
+  'Буду помнить об этих моментах',
+  'Секунды счастья',
+  'Выходные в кругу семьи',
+  'Иметь мягкое сердце в жестоком мире — это сила, а не слабость',
+  'Пятница — мое второе любимое слово',
+  'У вас никогда не заканчиваются вещи, которые могут пойти не так',
+  'Отдыхайте так, чтобы вы забывали брать телефон в руки',
+  'Будьте счастливы. Это сводит людей с ума',
+  'Будьте тем человеком, с которым вы хотите провести всю жизнь',
+  'Моя жизнь не идеальна, но это лучшее, что когда-либо случалось со мной'
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+const names = [
+  'Михаил',
+  'Александр',
+  'Максим',
+  'Марк',
+  'Артем',
+  'Лев',
+  'Матвей',
+  'Тимофей',
+  'София',
+  'Анна',
+  'Мария',
+  'Ева',
+  'Виктория',
+  'Полина',
+  'Варвара',
+  'Алиса',
+  'Василиса'
+];
+
+const surnames = [
+  'Ющенко',
+  'Тимошенко',
+  'Шнайдер',
+  'Шуткевич',
+  'Черных',
+  'Сухих',
+  'Чутких',
+  'Долгих',
+  'Бегун',
+  'Мельник',
+  'Соловей',
+  'Дюма',
+  'Ткаченко',
+  'Франко',
+  'Чуб',
+  'Худик',
+  'Трумпф'
+];
+
+export {descriptions, messages, names, surnames};
